Remove reversed time check from `combine_echodata` (#835)
* remove reversed time check for combined EchoData object, remove unnecessary test in test_echodata_combine and associated fixtures/parameters, include ek80 narrowband parameters in test_combine_echodata

* move reversed time check functions from combine.py to qc/api.py
