+++ echopype/echodata/combine.py
@@ -4,10 +4,8 @@ from warnings import warn
 
 import dask.distributed
 import fsspec
-import xarray as xr
 from dask.distributed import Client
 
-from ..qc import coerce_increasing_time, exist_reversed_time
 from ..utils.io import validate_output_path
 from ..utils.log import _init_logger
 from .echodata import EchoData
@@ -149,171 +147,6 @@ def check_echodatas_input(echodatas: List[EchoData]) -> Tuple[str, List[str]]:
     return sonar_model, echodata_filenames
 
 
-def check_and_correct_reversed_time(
-    combined_group: xr.Dataset, time_str: str, ed_group: str
-) -> Optional[xr.DataArray]:
-    """
-    Makes sure that the time coordinate ``time_str`` in
-    ``combined_group`` is in the correct order and corrects
-    it, if it is not. If coercion is necessary, the input
-    `combined_group` will be directly modified.
-
-    Parameters
-    ----------
-    combined_group : xr.Dataset
-        Dataset representing a combined EchoData group
-    time_str : str
-        Name of time coordinate to be checked and corrected
-    ed_group : str
-        Name of ``EchoData`` group name
-
-    Returns
-    -------
-    old_time : xr.DataArray or None
-        If correction is necessary, returns the time before
-        reversal correction, otherwise returns None
-
-    Warns
-    -----
-    UserWarning
-        If a time reversal is detected
-    """
-
-    if time_str in combined_group and exist_reversed_time(combined_group, time_str):
-
-        logger.warning(
-            f"{ed_group} {time_str} reversal detected; {time_str} will be corrected"  # noqa
-            " (see https://github.com/OSOceanAcoustics/echopype/pull/297)"
-        )
-        old_time = combined_group[time_str].copy()
-        coerce_increasing_time(combined_group, time_name=time_str)
-    else:
-        old_time = None
-
-    return old_time
-
-
-def create_old_time_array(group: str, old_time_in: xr.DataArray) -> xr.DataArray:
-    """
-    Creates an old time array with the appropriate values, name,
-    attributes, and encoding.
-
-    Parameters
-    ----------
-    group: str
-        The name of the ``EchoData`` group that contained
-        the old time
-    old_time_in: xr.DataArray
-        The uncorrected old time
-
-    Returns
-    -------
-    old_time_array: xr.DataArray
-        The newly created old time array
-    """
-
-    # make a copy, so we don't change the source array
-    old_time = old_time_in.copy()
-
-    # get name of old time and dim for Provenance group
-    ed_name = group.replace("-", "_").replace("/", "_").lower()
-    old_time_name = ed_name + "_old_" + old_time.name
-
-    old_time_name_dim = old_time_name + "_dim"
-
-    # construct old time attributes
-    attributes = old_time.attrs
-    attributes["comment"] = f"Uncorrected {old_time.name} from the combined group {group}."
-
-    # create old time array
-    old_time_array = xr.DataArray(
-        data=old_time.values, dims=[old_time_name_dim], attrs=attributes, name=old_time_name
-    )
-
-    # set encodings
-    old_time_array.encoding = old_time.encoding
-
-    return old_time_array
-
-
-def orchestrate_reverse_time_check(
-    ed_comb: EchoData, zarr_store: str, possible_time_dims: List[str], storage_options: dict
-) -> None:
-    """
-    Performs a reverse time check of all groups and
-    each time dimension within the group. If a reversed
-    time is found it will be corrected in ``ed_comb``,
-    updated in the zarr store, the old time will be
-    added to the ``Provenance`` group in ``ed_comb``,
-    the old time will be written to the zarr store,
-    and the attribute ``reversed_ping_times`` in the
-    ``Provenance`` group will be set to ``1``.
-
-    Parameters
-    ----------
-    ed_comb: EchoData
-        ``EchoData`` object that has been constructed from
-        combined ``EchoData`` objects
-    zarr_store: str
-        The zarr store containing the ``ed_comb`` data
-    possible_time_dims: list of str
-        All possible time dimensions that can occur within
-        ``ed_comb``, which should be checked
-    storage_options: dict
-        Additional keywords to pass to the filesystem class.
-
-    Notes
-    -----
-    If correction is necessary, ``ed_comb`` will be
-    directly modified.
-    """
-
-    # set Provenance attribute to zero in ed_comb
-    ed_comb["Provenance"].attrs["reversed_ping_times"] = 0
-
-    # set Provenance attribute to zero in zarr (Dataset needed for metadata creation)
-    only_attrs_ds = xr.Dataset(attrs=ed_comb["Provenance"].attrs)
-    only_attrs_ds.to_zarr(zarr_store, group="Provenance", mode="a", storage_options=storage_options)
-
-    for group in ed_comb.group_paths:
-
-        if group != "Platform/NMEA":
-            # Platform/NMEA is skipped because we found that the times which correspond to
-            # other non-GPS messages are often out of order and correcting them is not
-            # possible with the current implementation of _clean_ping_time in qc.api due
-            # to excessive recursion. There is also no obvious advantage in correcting
-            # the order of these timestamps.
-
-            # get all time dimensions of the group
-            ed_comb_time_dims = set(ed_comb[group].dims).intersection(possible_time_dims)
-
-            for time in ed_comb_time_dims:
-
-                old_time = check_and_correct_reversed_time(
-                    combined_group=ed_comb[group], time_str=time, ed_group=group
-                )
-
-                if old_time is not None:
-
-                    old_time_array = create_old_time_array(group, old_time)
-
-                    # put old times in Provenance and modify attribute
-                    ed_comb["Provenance"][old_time_array.name] = old_time_array
-                    ed_comb["Provenance"].attrs["reversed_ping_times"] = 1
-
-                    # save old time to zarr store
-                    old_time_ds = old_time_array.to_dataset()
-                    old_time_ds.attrs = ed_comb["Provenance"].attrs
-                    old_time_ds.to_zarr(
-                        zarr_store, group="Provenance", mode="a", storage_options=storage_options
-                    )
-
-                    # save corrected time to zarr store
-                    ed_comb[group][[time]].to_zarr(
-                        zarr_store, group=group, mode="r+", storage_options=storage_options
-                    )
-
-
 def combine_echodata(
     echodatas: List[EchoData] = None,
     zarr_path: Optional[str] = None,
@@ -379,12 +212,6 @@ def combine_echodata(
         - the keys ``date_created`` or ``conversion_time``
           do not have the same types
 
-    Warns
-    -----
-    UserWarning
-        If any time coordinate in a final combined group is not
-        in ascending order (see Notes below for more details).
-
     Notes
     -----
     * ``EchoData`` objects are combined by appending their groups individually to a zarr store.
@@ -392,10 +219,6 @@ def combine_echodata(
       combination will be stored in the ``Provenance`` group.
     * The instance attributes ``source_file`` and ``converted_raw_path`` of the combined
       ``EchoData`` object will be copied from the first ``EchoData`` object in the given list.
-    * If any time coordinate in a final combined group is not in ascending order, then it will
-      be corrected according to `#297 <https://github.com/OSOceanAcoustics/echopype/pull/297>`_.
-      Additionally, the uncorrected time coordinate will be stored in the ``Provenace`` group as
-      a variable and the ``Provenance`` attribute ``reversed_ping_times`` will be set to ``1``.
     * If no ``zarr_path`` is provided, the combined zarr file will be
       ``'temp_echopype_output/combined_echodatas.zarr'`` under the current working directory.
     * If no ``client`` is provided, then a client with a local scheduler will be used. The
@@ -420,7 +243,6 @@ def combine_echodata(
     >>>                                      zarr_path="path/to/combined.zarr",
     >>>                                      storage_options=my_storage_options)
     """
-    # TODO: change PR #297 reference to a link in our documentation
 
     # set flag specifying that a client was not created
     client_created = False
@@ -459,8 +281,6 @@ def combine_echodata(
         sonar_model=sonar_model,
         echodata_filenames=echodata_filenames,
     )
-
-    orchestrate_reverse_time_check(ed_comb, zarr_path, comb.possible_time_dims, storage_options)
 
     if client_created:
         # close client
