+++ echopype/tests/echodata/test_echodata_combine.py
@@ -7,7 +7,6 @@ import xarray as xr
 
 import echopype
 from echopype.utils.coding import DEFAULT_ENCODINGS
-from echopype.qc import exist_reversed_time
 
 import tempfile
 from dask.distributed import Client
@@ -19,16 +18,6 @@ def ek60_test_data(test_path):
         ("ncei-wcsd", "Summer2017-D20170620-T011027.raw"),
         ("ncei-wcsd", "Summer2017-D20170620-T014302.raw"),
         ("ncei-wcsd", "Summer2017-D20170620-T021537.raw"),
-    ]
-    return [test_path["EK60"].joinpath(*f) for f in files]
-
-
-@pytest.fixture
-def ek60_reversed_ping_time_test_data(test_path):
-    files = [
-        ("ncei-wcsd", "Summer2017-D20170719-T203615.raw"),
-        ("ncei-wcsd", "Summer2017-D20170719-T205415.raw"),
-        ("ncei-wcsd", "Summer2017-D20170719-T211347.raw"),
     ]
     return [test_path["EK60"].joinpath(*f) for f in files]
 
@@ -108,11 +97,6 @@ def azfp_test_xml(test_path):
             "files": "ek60_diff_range_sample_test_data"
         },
         {
-            "sonar_model": "EK60",
-            "xml_file": None,
-            "files": "ek60_reversed_ping_time_test_data"
-        },
-        {
             "sonar_model": "AZFP",
             "xml_file": "azfp_test_xml",
             "files": "azfp_test_data"
@@ -128,7 +112,7 @@ def azfp_test_xml(test_path):
             "files": "ek80_narrowband_diff_range_sample_test_data"
         }
     ],
-    ids=["ek60", "ek60_diff_range_sample", "ek60_reversed_time", "azfp",
+    ids=["ek60", "ek60_diff_range_sample", "azfp",
          "ek80_bb_same_range_sample", "ek80_nb_diff_range_sample"]
 )
 def raw_datasets(request):
@@ -154,10 +138,6 @@ def test_combine_echodata(raw_datasets):
         xml_file,
         param_id,
     ) = raw_datasets
-
-    if param_id == "ek80_nb_diff_range_sample":
-        pytest.xfail("The files in ek80_nb_diff_range_sample cause an error when correcting a reversed time. "
-                     "Once this is fixed, these files should be ran.")
 
     eds = [echopype.open_raw(file, sonar_model, xml_file) for file in files]
 
@@ -185,107 +165,50 @@ def test_combine_echodata(raw_datasets):
     # add dimension for Provenance group
     all_drop_dims.append("echodata_filename")
 
-    # add dimension for reversed time that will show up in Provenance group
-    all_drop_dims.append("platform_old_time1_dim")
-
     for group_name in combined.group_paths:
 
-        # ignore the Platform group for ek60_reversed_time parameters (these are checked elsewhere)
-        if (param_id != "ek60_reversed_time") or (group_name != "Platform"):
+        # get all Datasets to be combined
+        combined_group: xr.Dataset = combined[group_name]
+        eds_groups = [
+            ed[group_name]
+            for ed in eds
+            if ed[group_name] is not None
+        ]
 
-            # get all Datasets to be combined
-            combined_group: xr.Dataset = combined[group_name]
-            eds_groups = [
-                ed[group_name]
-                for ed in eds
-                if ed[group_name] is not None
-            ]
-
-            # all grp dimensions that are in all_drop_dims
-            if combined_group is None:
-                grp_drop_dims = []
-                concat_dims = []
-            else:
-                grp_drop_dims = list(set(combined_group.dims).intersection(set(all_drop_dims)))
-                concat_dims = list(set(combined_group.dims).intersection(append_dims))
-
-            # concat all Datasets along each concat dimension
-            diff_concats = []
-            for dim in concat_dims:
-
-                drop_dims = [c_dim for c_dim in concat_dims if c_dim != dim]
-
-                diff_concats.append(xr.concat([ed_subset.drop_dims(drop_dims) for ed_subset in eds_groups], dim=dim,
-                                    coords="minimal", data_vars="minimal"))
-
-            if len(diff_concats) < 1:
-                test_ds = eds_groups[0]  # needed for groups that do not have append dims
-            else:
-                # create the full combined Dataset
-                test_ds = xr.merge(diff_concats, compat="override")
-
-                # correctly set filenames values for constructed combined Dataset
-                if "filenames" in test_ds:
-                    test_ds.filenames.values[:] = np.arange(len(test_ds.filenames), dtype=int)
-
-                # correctly modify Provenance attributes so we can do a direct compare
-                if group_name == "Provenance":
-
-                    if param_id == "ek60_reversed_time":
-                        test_ds.attrs["reversed_ping_times"] = 1
-                    else:
-                        test_ds.attrs["reversed_ping_times"] = 0
-
-                    del test_ds.attrs["conversion_time"]
-                    del combined_group.attrs["conversion_time"]
-
-            if (combined_group is not None) and (test_ds is not None):
-                assert test_ds.identical(combined_group.drop_dims(grp_drop_dims))
-
-    temp_zarr_dir.cleanup()
-
-    # close client
-    client.close()
-
-
-def test_ping_time_reversal(ek60_reversed_ping_time_test_data):
-
-    eds = [
-        echopype.open_raw(file, "EK60")
-        for file in ek60_reversed_ping_time_test_data
-    ]
-
-    # create temporary directory for zarr store
-    temp_zarr_dir = tempfile.TemporaryDirectory()
-    zarr_file_name = temp_zarr_dir.name + "/combined_echodatas.zarr"
-
-    # create dask client
-    client = Client()
-
-    combined = echopype.combine_echodata(eds, zarr_file_name, client=client)
-
-    for group_name, value in combined.group_map.items():
-        if value['ep_group'] is None:
-            combined_group: xr.Dataset = combined['Top-level']
+        # all grp dimensions that are in all_drop_dims
+        if combined_group is None:
+            grp_drop_dims = []
+            concat_dims = []
         else:
-            combined_group: xr.Dataset = combined[value['ep_group']]
+            grp_drop_dims = list(set(combined_group.dims).intersection(set(all_drop_dims)))
+            concat_dims = list(set(combined_group.dims).intersection(append_dims))
 
-        if combined_group is not None:
-            if "ping_time" in combined_group and group_name != "provenance":
-                assert not exist_reversed_time(combined_group, "ping_time")
-            if "old_ping_time" in combined_group:
-                assert exist_reversed_time(combined_group, "old_ping_time")
-            if "time1" in combined_group and group_name not in (
-                "provenance",
-                "nmea",
-            ):
-                assert not exist_reversed_time(combined_group, "time1")
-            if "old_time1" in combined_group:
-                assert exist_reversed_time(combined_group, "old_time1")
-            if "time2" in combined_group and group_name != "provenance":
-                assert not exist_reversed_time(combined_group, "time2")
-            if "old_time2" in combined_group:
-                assert exist_reversed_time(combined_group, "old_time2")
+        # concat all Datasets along each concat dimension
+        diff_concats = []
+        for dim in concat_dims:
+
+            drop_dims = [c_dim for c_dim in concat_dims if c_dim != dim]
+
+            diff_concats.append(xr.concat([ed_subset.drop_dims(drop_dims) for ed_subset in eds_groups], dim=dim,
+                                coords="minimal", data_vars="minimal"))
+
+        if len(diff_concats) < 1:
+            test_ds = eds_groups[0]  # needed for groups that do not have append dims
+        else:
+            # create the full combined Dataset
+            test_ds = xr.merge(diff_concats, compat="override")
+
+            # correctly set filenames values for constructed combined Dataset
+            if "filenames" in test_ds:
+                test_ds.filenames.values[:] = np.arange(len(test_ds.filenames), dtype=int)
+
+            # correctly modify Provenance attributes, so we can do a direct compare
+            if group_name == "Provenance":
+                del test_ds.attrs["conversion_time"]
+                del combined_group.attrs["conversion_time"]
+
+        if (combined_group is not None) and (test_ds is not None):
+            assert test_ds.identical(combined_group.drop_dims(grp_drop_dims))
 
     temp_zarr_dir.cleanup()
